+++ src/components/FullScreenDialog.tsx
@@ -1,12 +1,40 @@
-/* eslint-disable require-jsdoc */
 import React from 'react'
 import { AppBar, Dialog, IconButton, Toolbar, Typography } from '@material-ui/core'
 import CloseIcon from '@material-ui/icons/Close'
 import SaveAlt from '@material-ui/icons/SaveAlt'
-import { Transition, useStyles } from './app'
-import { DummyDrawerRight } from './ResponsiveDrawer'
-import { FullScreenprops } from './Interface'
+import { makeStyles } from '@material-ui/core/styles'
+import { Transition } from '../app'
+import { SelectedImage } from '../Interface'
+import ResponsiveDrawer from './ResponsiveDrawer'
 
+interface FullScreenprops {
+  openedImg: SelectedImage
+  isopen: boolean
+  closeFunction: () => void
+  steppingFunction: (imageIndex: number, openInfoTab: boolean) => void
+  imgList: object[]
+}
+
+export const useStyles = makeStyles(theme => ({
+  root: {
+    display: 'flex',
+    flexWrap: 'wrap',
+    justifyContent: 'space-around',
+    overflow: 'visible',
+    backgroundColor: theme.palette.background.paper,
+  },
+  title: {
+    marginLeft: theme.spacing(2),
+    flex: 1,
+  },
+  appBar: {
+    width: '100%',
+    marginRight: '0',
+  },
+}))
+/**
+ * Display the details view
+ */
 export default function FullScreenDialog(props: FullScreenprops) {
   const classes = useStyles()
   return (
@@ -25,8 +53,7 @@ export default function FullScreenDialog(props: FullScreenprops) {
             </IconButton>
           </Toolbar>
         </AppBar>
-        {/* <ResponsiveDrawer openededImg={props.openedImg} /> */}
-        <DummyDrawerRight
+        <ResponsiveDrawer
           steppingFunction={props.steppingFunction}
           imageListLenght={props.imgList.length}
           imageInfo={props.openedImg}
